refactor(ios): decompose SettingsView into focused section files
Split SettingsView.swift (765 lines) into 6 section files under
Settings/ directory. SettingsView.swift is now a ~90-line coordinator.

New files:
- Settings/ConnectionSettingsSection.swift — API key + daemon config
- Settings/IntegrationsSection.swift — integration list
- Settings/TrustRulesSection.swift — trust rules + form
- Settings/SchedulesSection.swift — scheduled tasks
- Settings/RemindersSection.swift — reminders
- Settings/PermissionRowView.swift — permission row

Pure file reorganization — no logic or behavior changes. Each section
manages its own state and daemon callbacks independently.
