+++ clients/ios/Views/Settings/SchedulesSection.swift
@@ -1,0 +1,107 @@
+#if canImport(UIKit)
+import SwiftUI
+import VellumAssistantShared
+
+struct SchedulesSection: View {
+    @EnvironmentObject var clientProvider: ClientProvider
+    @State private var schedules: [ScheduleItem] = []
+    @State private var loading = false
+
+    var body: some View {
+        Section("Scheduled Tasks") {
+            if loading {
+                HStack {
+                    Spacer()
+                    ProgressView()
+                    Spacer()
+                }
+            } else if schedules.isEmpty {
+                Text("No scheduled tasks")
+                    .foregroundStyle(.secondary)
+                    .font(.caption)
+            } else {
+                ForEach(schedules, id: \.id) { schedule in
+                    scheduleRow(schedule)
+                }
+                .onDelete { indexSet in
+                    let schedulesToDelete = indexSet.map { schedules[$0] }
+                    for schedule in schedulesToDelete {
+                        deleteSchedule(schedule.id)
+                    }
+                }
+            }
+        }
+        .onAppear { loadSchedules() }
+        .onDisappear {
+            if let daemon = clientProvider.client as? DaemonClient {
+                daemon.onSchedulesListResponse = nil
+            }
+        }
+    }
+
+    @ViewBuilder
+    private func scheduleRow(_ schedule: ScheduleItem) -> some View {
+        HStack {
+            VStack(alignment: .leading, spacing: 2) {
+                Text(schedule.name)
+                    .font(.body)
+                Text(schedule.description)
+                    .font(.caption)
+                    .foregroundStyle(.secondary)
+                    .lineLimit(2)
+                HStack(spacing: 4) {
+                    Text(schedule.cronExpression)
+                        .font(.caption2)
+                        .foregroundStyle(.tertiary)
+                    if schedule.enabled, schedule.nextRunAt > 0, let nextRun = formatTimestamp(schedule.nextRunAt) {
+                        Text("Next: \(nextRun)")
+                            .font(.caption2)
+                            .foregroundStyle(.tertiary)
+                    }
+                }
+            }
+            Spacer()
+            Toggle("", isOn: Binding(
+                get: { schedule.enabled },
+                set: { newValue in toggleSchedule(schedule.id, enabled: newValue) }
+            ))
+            .labelsHidden()
+        }
+    }
+
+    private func loadSchedules() {
+        guard let daemon = clientProvider.client as? DaemonClient else { return }
+        loading = true
+        daemon.onSchedulesListResponse = { items in
+            schedules = items
+            loading = false
+        }
+        do {
+            try daemon.sendListSchedules()
+        } catch {
+            loading = false
+        }
+    }
+
+    private func toggleSchedule(_ id: String, enabled: Bool) {
+        guard let daemon = clientProvider.client as? DaemonClient else { return }
+        try? daemon.sendToggleSchedule(id: id, enabled: enabled)
+        if schedules.firstIndex(where: { $0.id == id }) != nil {
+            loadSchedules()
+        }
+    }
+
+    private func deleteSchedule(_ id: String) {
+        guard let daemon = clientProvider.client as? DaemonClient else { return }
+        try? daemon.sendRemoveSchedule(id: id)
+        schedules.removeAll { $0.id == id }
+    }
+
+    private func formatTimestamp(_ ms: Int) -> String? {
+        let date = Date(timeIntervalSince1970: TimeInterval(ms) / 1000.0)
+        let formatter = RelativeDateTimeFormatter()
+        formatter.unitsStyle = .abbreviated
+        return formatter.localizedString(for: date, relativeTo: Date())
+    }
+}
+#endif
